Update build files

+++ build/server.py
@@ -6,13 +6,16 @@ import os
 print("[+] Ready to set up directorys, any existing talkative infuastructure shall be overwritten...")
 x = str(input("Press enter to begin: "))
 
-paths = ["/usr/share/Talkative","/usr/share/Talkative/Server/Users","/usr/share/Talkative/Server/Users/DefaultUser1","/usr/share/Talkative/Server/Chats","/usr/share/Talkative/Server/Src"]
+paths = ["/usr/share/Talkative","/usr/share/Talkative/server/users","/usr/share/Talkative/server/users/DefaultUser1","/usr/share/Talkative/server/Chats","/usr/share/Talkative/server/src"]
 
 for p in paths:
     if os.path.exists(p):
-        print(f"[+] Found existing directory: {path}, overwriting")
-        os.system(f"rm -r {p}")
-        os.system(f"mkdir {p}")
+        if paths.index(p) == 0:
+            continue
+        else:
+            print(f"[+] Found existing directory: {path}, overwriting")
+            os.system(f"rm -r {p}")
+            os.system(f"mkdir {p}")
     else:
         os.system(f"mkdir {p}")
 
@@ -23,7 +26,7 @@ print("[+] Completed directory setup, ready to begin file setup")
 x = str(input("Press enter to begin: "))
 
 
-os.system("mv ~/Downloads/Talkative/Src/Server/* /usr/share/Talkative/Server/Src")
-os.system("touch /usr/share/Talkative/Src/Server/Users/DefaultUser1/user_chats.txt")
+os.system("mv ~/Downloads/Talkative/src/server/* /usr/share/Talkative/server/src")
+os.system("touch /usr/share/Talkative/src/server/users/DefaultUser1/user_chats.txt")
